14513:   Fix incorrect variable type in SPILength constructor.

+++ src/style-internal.h
@@ -342,7 +342,7 @@ public:
           computed(0)
     {}
 
-    SPILength( Glib::ustring const &name, unsigned value = 0 )
+    SPILength( Glib::ustring const &name, float value = 0 )
         : SPIBase( name ),
           unit(SP_CSS_UNIT_NONE),
           value(value),
